requirements.txt
# GUI Framework
customtkinter>=5.2.0
tkinter-tooltip>=2.1.0

# PDF Generation
reportlab>=4.0.4
pillow>=10.0.0

# Data Handling
pandas>=2.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Security
cryptography>=41.0.0
bcrypt>=4.0.0

# Image Processing
Pillow>=10.0.0

# Configuration
pyyaml>=6.0
configparser>=5.3.0

# Validation
cerberus>=1.3.4
marshmallow>=3.20.0

# Logging
structlog>=23.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# QR Code Generation
qrcode>=7.4.2

# Rich Text
markdown>=3.4.0

# File Watching
watchdog>=3.0.0

# Audio Support (Optional - for pronunciation practice)
# Install with: pip install pyaudio sounddevice soundfile pygame
# pyaudio>=0.2.11
# sounddevice>=0.4.6
# soundfile>=0.12.1
# pygame>=2.5.0
# numpy>=1.24.0

# Document Import (optional)
# python-docx allows importing .docx files: pip install python-docx
# python-docx>=0.8.11